валидация перенесена из репозитория в сервис для пользователя
