+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -2,7 +2,9 @@ package ru.practicum.shareit.user;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.DuplicatedDataException;
 import ru.practicum.shareit.exception.NotFoundException;
+import ru.practicum.shareit.exception.ValidationException;
 import ru.practicum.shareit.user.dto.UserDto;
 
 import java.util.List;
@@ -30,17 +32,46 @@ public class UserService {
     }
 
     public UserDto create(User user) {
+        checkEmail(user);
+        checkName(user);
         return UserMapper.mapToUserDto(userRepository.create(user));
     }
 
     public UserDto update(User newUser) {
+        checkUserExists(newUser.getId());
+        if (newUser.getEmail() != null) {
+            checkEmail(newUser);
+        }
         return UserMapper.mapToUserDto(userRepository.update(newUser));
     }
 
     public UserDto delete(Long id) {
+        checkUserExists(id);
         return userRepository.delete(id)
                 .map(UserMapper::mapToUserDto)
                 .orElseThrow(() -> new NotFoundException(String.format("Пользователь с id=%d не найден", id)));
+    }
+
+    private void checkUserExists(long userId) {
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new NotFoundException(String.format("Пользователь с id=%d не найден", userId));
+        }
+    }
+
+    private void checkName(User user) {
+        if (user.getName() == null || user.getName().isBlank()) {
+            log.error("Имя пользователя не указано");
+            throw new ValidationException("Имя должно быть указано");
+        }
+    }
+
+    private void checkEmail(User user) {
+        for (User value : userRepository.users.values()) {
+            if (user.getEmail().equals(value.getEmail())) {
+                log.error("Попытка занять уже используемый email");
+                throw new DuplicatedDataException("Этот email уже используется");
+            }
+        }
     }
 }
